Update workflow name and Docker image tag; Remove paymentStatus from OrderQueueDTO
The workflow name in the GitHub Actions configuration is updated from "Testes PR" to "Deployment". The names and tags of the Docker image built and pushed to AWS ECR are also modified. Moreover, the `paymentStatus` field has been removed from the `OrderQueueDTO` class and all usage instances in related converter classes.