Add Jacoco code coverage checks to GitHub Actions workflow
This commit adds steps to the GitHub Actions PR workflow in order to generate and upload Jacoco coverage reports for each run. It also enforces a minimum coverage of 80%. The SonarQube plugin was removed and adjustments were made to the `build.gradle` file to ensure correct calculation and reporting of code coverage statistics.